+++ backend/utils.py
@@ -548,7 +548,22 @@ mods_props = {
         }
     }
 }
-            
+
+# combine scores for each matching criterion to overall score
+
+def matches_to_score(matches):
+    
+    weights = [
+        10,  # ref:IFOPT
+        2,  # ref
+        2,  # local_ref
+        1,  # ref_name
+        1,  # name
+        1,  # layer
+        2  # level
+    ]
+    return sum([w * m for w, m in zip(weights, matches)])
+
 
 # function for assigning nodes to nodes or areas via neighborhood relations
 
@@ -585,17 +600,12 @@ def get_nearby_nodes(nodes, objects, col_prefix, radius, mods_func, score_func):
         osm_obj = objects.loc[obj_id, 'obj']
         for node_id in node_ids:
             node_info = objects.loc[obj_id, infos_col][node_id]
+            matches = score_func(osm_obj, nodes.loc[node_id, 'obj'])
             node_info['ref:IFOPT_match'], node_info['ref_match'], \
             node_info['local_ref_match'], node_info['ref_name_match'], \
             node_info['name_match'], node_info['layer_match'], node_info['level_match'] \
-                = score_func(osm_obj, nodes.loc[node_id, 'obj'])
-            node_info['score'] = 10 * node_info['ref:IFOPT_match'] \
-                                 + 2 * node_info['ref_match'] \
-                                 + 2 * node_info['local_ref_match'] \
-                                 + 1 * node_info['ref_name_match'] \
-                                 + 1 * node_info['name_match'] \
-                                 + 1 * node_info['layer_match'] \
-                                 + 2 * node_info['level_match']
+                = matches
+            node_info['score'] = matches_to_score(matches)
 
         # adjust scores by distance (add 0...1/2 for distance max...0)
         # note: this adjustment by distance only influences the ordering of